Add signAgreement action
