+++ src/components/ModalFlows/SignAgreementScreens/SignAgreementScreens.js
@@ -1,34 +1,24 @@
 import React, { useMemo, useState, useCallback } from 'react'
 import ModalFlowBase from '../ModalFlowBase'
 import SignOverview from './SignOverview'
-import { useMounted } from '../../../hooks/useMounted'
-import { useWallet } from '../../../providers/Wallet'
-import { useOrgApps } from '../../../providers/OrgApps'
 import { useAgreementState } from '../../../providers/AgreementState'
+import { useActions } from '../../../hooks/useActions'
 
 function SignAgreementScreens() {
-  const mounted = useMounted()
-  const { account } = useWallet()
-  const { agreementApp } = useOrgApps()
   const {
     agreement: { versionId },
   } = useAgreementState()
+  const { signAgreement } = useActions()
   const [transactions, setTransactions] = useState([])
 
   const getTransactions = useCallback(
     async (onComplete) => {
-      try {
-        const { transactions } = await agreementApp.sign(account, versionId)
-
-        if (mounted()) {
-          setTransactions(transactions)
-          onComplete()
-        }
-      } catch (err) {
-        console.error(err)
-      }
+      await signAgreement(versionId, (intent) => {
+        setTransactions(intent.transactions)
+        onComplete()
+      })
     },
-    [account, agreementApp, versionId, mounted]
+    [signAgreement, versionId]
   )
 
   const screens = useMemo(
